# docs/swagger.yaml
basePath: /api/v1
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Address:
    properties:
      city:
        description: City name
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      number:
        description: Building number
        type: string
      owner_id:
        description: Owner's ID
        type: integer
      owner_type:
        description: Owner's type
        type: string
      postal:
        description: Postal code
        type: string
      street:
        description: Street name
        type: string
      updatedAt:
        type: string
    type: object
  model.Class:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      is_active:
        description: Whether the class is active
        type: boolean
      name:
        description: Name of the class
        type: string
      school_id:
        description: ID of the school the class belongs to
        type: integer
      students:
        description: List of students in the class
        items:
          $ref: '#/definitions/model.Student'
        type: array
      teacher_id:
        description: ID of the teacher for the class
        type: integer
      updatedAt:
        type: string
      year:
        description: Year of the class
        type: string
    type: object
  model.Exam:
    properties:
      createdAt:
        type: string
      date:
        description: Date of the exam
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: Description of the exam
        type: string
      exam_type:
        allOf:
        - $ref: '#/definitions/model.ExamType'
        description: Exam type
      exam_type_id:
        description: ID of the exam type
        type: integer
      id:
        type: integer
      max_points:
        description: Maximum points for the exam
        type: integer
      name:
        description: Name of the exam
        type: string
      scores:
        allOf:
        - $ref: '#/definitions/model.Score'
        description: Scores for the exam
      updatedAt:
        type: string
    type: object
  model.ExamType:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        description: Name is the name of the exam type.
        type: string
      updatedAt:
        type: string
    type: object
  model.Grade:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        description: Name is the name of the grade.
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.Score:
    properties:
      comment:
        description: Comment represents any additional comments about the score.
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      exam_id:
        description: ExamID represents the ID of the associated exam.
        type: integer
      grade:
        allOf:
        - $ref: '#/definitions/model.Grade'
        description: Grade represents the associated grade.
      grade_id:
        description: GradeID represents the ID of the associated grade.
        type: integer
      id:
        type: integer
      points:
        description: Points represents the number of points scored by the student.
        type: integer
      student_id:
        description: StudentID represents the ID of the student.
        type: integer
      updatedAt:
        type: string
    type: object
  model.Student:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/model.Address'
        description: Address represents the address of the student
      class_id:
        description: ClassID represents the ID of the class the student belongs to
        type: integer
      createdAt:
        type: string
      date_of_birth:
        description: DateOfBirth represents the date of birth of the student
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      nachname:
        description: Nachname represents the last name of the student
        type: string
      updatedAt:
        type: string
      vorname:
        description: Vorname represents the first name of the student
        type: string
    type: object
  model.Subject:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        description: Name field for the subject, marked as not null in the database
          and exposed as "name" in JSON
        type: string
      updatedAt:
        type: string
    type: object
  model.SubjectUserClassExam:
    properties:
      class:
        allOf:
        - $ref: '#/definitions/model.Class'
        description: Class is the class associated with this record.
      class_id:
        description: ClassID is the ID of the class.
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      exam:
        allOf:
        - $ref: '#/definitions/model.Exam'
        description: Exam is the exam associated with this record.
      exam_id:
        description: ExamID is the ID of the exam.
        type: integer
      id:
        type: integer
      subject:
        allOf:
        - $ref: '#/definitions/model.Subject'
        description: Subject is the subject associated with this record.
      subject_id:
        description: SubjectID is the ID of the subject.
        type: integer
      updatedAt:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: User is the user associated with this record.
      user_id:
        description: UserID is the ID of the user.
        type: integer
    type: object
  model.User:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/model.Address'
        description: Address is the user's address.
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        description: Email is the user's email address.
        type: string
      id:
        type: integer
      is_admin:
        description: IsAdmin indicates whether the user is an administrator.
        type: boolean
      nachname:
        description: Nachname is the user's last name.
        type: string
      password:
        description: Password is the user's password.
        type: string
      school_id:
        description: SchoolID is the ID of the user's school.
        type: integer
      updatedAt:
        type: string
      username:
        description: Username is the user's username.
        type: string
      vorname:
        description: Vorname is the user's first name.
        type: string
    type: object
  routes.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:1323
info:
  contact:
    email: patrykhegenberg@gmail.com
    name: API support
    url: http://swagger.io/support
  license:
    name: Apache 2.0
    url: http://apache.com/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms
  title: Notenverwaltung API
  version: 0.0.8
paths:
...
  /subjects:
    get:
      consumes:
      - application/json
      description: get all subjects from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subject'
            type: array
        "500":
          description: Fehler beim Abrufen der Subjecten
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get all subjects
      tags:
      - subject
    post:
      consumes:
      - application/json
      description: create a new subject
      parameters:
      - description: Subject object that needs to be added
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/model.Subject'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subject'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Erstellen des Subjecten
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: create subject
      tags:
      - subject
  /subjects/:id:
    delete:
      consumes:
      - application/json
      description: delete an existing subject
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Ungültige Subjecten-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Löschen des Subjecten
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: delete subject
      tags:
      - subject
    get:
      consumes:
      - application/json
      description: get one subject from db by ID.
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subject'
        "400":
          description: Ungültige Subjecten-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Subject nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get one subject by id
      tags:
      - subject
    put:
      consumes:
      - application/json
      description: update an existing subject
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated subject object
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/model.Subject'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subject'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Subject nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: update subject
      tags:
      - subject
  /sucs:
    get:
      consumes:
      - application/json
      description: get all subject user class exams from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SubjectUserClassExam'
            type: array
        "500":
          description: Fehler beim Abrufen der SubjectUserClassExamen
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get all subject user class exams
      tags:
      - subjectuserclassexam
    post:
      consumes:
      - application/json
      description: create a new subject user class exam
      parameters:
      - description: SubjectUserClassExam object that needs to be added
        in: body
        name: subjectUserClassExam
        required: true
        schema:
          $ref: '#/definitions/model.SubjectUserClassExam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.SubjectUserClassExam'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Erstellen des SubjectUserClassExamen
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: create subject user class exam
      tags:
      - subjectuserclassexam
  /sucs/:id:
    delete:
      consumes:
      - application/json
      description: delete an existing subject user class exam
      parameters:
      - description: SubjectUserClassExam ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Ungültige SubjectUserClassExamen-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Löschen des SubjectUserClassExamen
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: delete subject user class exam
      tags:
      - subjectuserclassexam
    get:
      consumes:
      - application/json
      description: get one subject user class exam from db by ID.
      parameters:
      - description: SubjectUserClassExam ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SubjectUserClassExam'
        "400":
          description: Ungültige SubjectUserClassExamen-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: SubjectUserClassExam nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get one subject user class exam by id
      tags:
      - subjectuserclassexam
    put:
      consumes:
      - application/json
      description: update an existing subject user class exam
      parameters:
      - description: SubjectUserClassExam ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated subject user class exam object
        in: body
        name: subjectUserClassExam
        required: true
        schema:
          $ref: '#/definitions/model.SubjectUserClassExam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SubjectUserClassExam'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: SubjectUserClassExam nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: update subject user class exam
      tags:
      - subjectuserclassexam
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"